+++ include/display_utils.h
@@ -17,5 +17,6 @@ void drawStatusBar();
 void drawWifiIcon(int x, int y, long rssi);
 void drawAPIcon(int x, int y);
 void drawBatteryIcon(int x, int y, float voltage);
+float getBatteryVoltage();
 
 #endif